CDynamicGrid<> now uses `double`  for better accuracy

+++ libs/maps/include/mrpt/maps/CHeightGridMap2D.h
@@ -76,11 +76,9 @@ namespace mrpt
 			/** Constructor */
 			CHeightGridMap2D(
 				TMapRepresentation	mapType = mrSimpleAverage,
-				float				x_min = -2,
-				float				x_max = 2,
-				float				y_min = -2,
-				float				y_max = 2,
-				float				resolution = 0.1
+				double x_min = -2, double x_max = 2,
+				double y_min = -2, double y_max = 2,
+				double resolution = 0.1
 				);
 
 			 bool isEmpty() const MRPT_OVERRIDE; //!< Returns true if the map is empty/no observation has been inserted.
@@ -137,7 +135,7 @@ namespace mrpt
 			double internal_computeObservationLikelihood( const mrpt::obs::CObservation *obs, const mrpt::poses::CPose3D &takenFrom ) MRPT_OVERRIDE;
 
 			MAP_DEFINITION_START(CHeightGridMap2D,MAPS_IMPEXP)
-				float	min_x,max_x,min_y,max_y,resolution;	//!< See CHeightGridMap2D::CHeightGridMap2D
+				double min_x,max_x,min_y,max_y,resolution;	//!< See CHeightGridMap2D::CHeightGridMap2D
 				mrpt::maps::CHeightGridMap2D::TMapRepresentation	mapType;	//!< The kind of map representation (see CHeightGridMap2D::CHeightGridMap2D)
 				mrpt::maps::CHeightGridMap2D::TInsertionOptions	insertionOpts;
 			MAP_DEFINITION_END(CHeightGridMap2D,MAPS_IMPEXP)
