+++ libs/maps/include/mrpt/maps/CReflectivityGridMap2D.h
@@ -61,18 +61,10 @@ namespace mrpt
 				return m_logodd_lut.l2p(c);
 			}
 
-			/** Constructor
-			  */
-			CReflectivityGridMap2D(
-				float				x_min = -2,
-				float				x_max = 2,
-				float				y_min = -2,
-				float				y_max = 2,
-				float				resolution = 0.1
-				);
+			/** Constructor  */
+			CReflectivityGridMap2D(double x_min = -2, double x_max = 2, double y_min = -2, double y_max = 2,double resolution = 0.1);
 
-			 /** Returns true if the map is empty/no observation has been inserted.
-			   */
+			 /** Returns true if the map is empty/no observation has been inserted. */
 			 bool isEmpty() const MRPT_OVERRIDE;
 
 
@@ -110,7 +102,7 @@ namespace mrpt
 			double internal_computeObservationLikelihood( const mrpt::obs::CObservation *obs, const mrpt::poses::CPose3D &takenFrom )  MRPT_OVERRIDE;
 
 			MAP_DEFINITION_START(CReflectivityGridMap2D,MAPS_IMPEXP)
-				float	min_x,max_x,min_y,max_y,resolution;	//!< See CReflectivityGridMap2DOptions::CReflectivityGridMap2DOptions
+				double min_x,max_x,min_y,max_y,resolution;	//!< See CReflectivityGridMap2DOptions::CReflectivityGridMap2DOptions
 				mrpt::maps::CReflectivityGridMap2D::TInsertionOptions	insertionOpts;
 			MAP_DEFINITION_END(CReflectivityGridMap2D,MAPS_IMPEXP)
 		};
